Implemented file permissions and owner tracking
