--- file_processing/analytics.csv
@@ -1,31 +1,0 @@
-Extension,Size,Count
-.abcd,20,1
-.aiff,51142728,1
-.csv,17562307,4
-.docx,97953,6
-.flac,12938126,1
-.gif,4332710,6
-.heic,224590,2
-.heif,224590,2
-.html,503157,3
-.jpeg,8800,1
-.jpg,3949606,5
-.json,2098336,3
-.mp3,27514263,3
-.mp4,10808772,1
-.msg,563200,4
-.ogg,1523997,1
-.pdf,1627953,11
-.png,17004074,7
-.pptx,100475886,8
-.py,384391,30
-.pyc,111514,28
-.rtf,309288,3
-.tif,6113650,5
-.tiff,93494,4
-.txt,223290,11
-.wav,110496300,3
-.xlsx,177368,6
-.xml,13525,3
-.xyz,17,1
-.zip,192784,4
