+++ tests/resources/sample_reports/analytics.csv
@@ -1,0 +1,34 @@
+Extension,Size (MB),Count
+,0.046164,2
+.abcd,2e-05,1
+.aiff,51.142728,1
+.csv,17.622422999999998,6
+.docx,0.09795300000000001,6
+.flac,12.938126,1
+.gif,4.33271,6
+.heic,0.22459,2
+.heif,0.22459,2
+.html,0.503157,3
+.ipynb,0.118149,2
+.jpeg,0.0088,1
+.jpg,3.9496059999999997,5
+.json,2.0983359999999998,3
+.md,0.005835,2
+.mp3,27.514263000000003,3
+.mp4,10.808772,1
+.msg,0.5631999999999999,4
+.ogg,1.523997,1
+.pdf,1.6279529999999998,11
+.png,17.004074,7
+.pptx,100.47588599999999,8
+.py,0.4541880000000001,62
+.rtf,0.309288,3
+.tif,6.11365,5
+.tiff,0.093494,4
+.toml,0.000161,1
+.txt,0.22395700000000004,12
+.wav,110.49629999999999,3
+.xlsx,0.17736800000000003,6
+.xml,0.013524999999999999,3
+.xyz,1.7e-05,1
+.zip,0.192784,4
